% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{addEdg}
\alias{addEdg}
\title{Add an edge in the in silico system.}
\usage{
addEdg(insilicosystem, regID, tarID, targetreaction, regsign,
  kinetics = NULL)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{regID}{Integer. The ID of the regulator gene.}

\item{tarID}{Integer. The ID of the target gene.}

\item{targetreaction}{The type of regulation exerted ("TC", "TL", "RD", "PD" or "PTM").}

\item{regsign}{The sign of the regulation: either "1" (positive regulation) or "-1" (negative regulation).}

\item{kinetics}{Optional: list of kinetics parameters of the reaction. If not provided, they will be sampled
according to the distributions specified in \code{sysargs} (parameters of the simulation - an attribute of the in silico system).}
}
\value{
The modified in silico system.
}
\description{
Add an edge in the in silico system between specified genes.
}
