% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{addEdge}
\alias{addEdge}
\title{Adds an edge in the in silico system.}
\usage{
addEdge(insilicosystem, regID, tarID, regsign = NULL,
  kinetics = list())
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{regID}{Integer. The ID of the regulator gene.}

\item{tarID}{Integer. The ID of the target gene.}

\item{regsign}{The sign of the regulation: either "1" (positive regulation) or "-1" (negative regulation). If none provided,
will be randomly chosen according to the parameters provided in sysargs (see \code{\link{insilicosystemargs}}).}

\item{kinetics}{Optional: named vector of kinetics parameters of the reaction. If none provided,
will be randomly chosen according to the parameters provided in sysargs (see \code{\link{insilicosystemargs}}).}
}
\value{
The modified in silico system.
}
\description{
Adds an edge in the in silico system between specified genes.
}
