% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{addGene}
\alias{addGene}
\title{Adds a gene in the in silico system.}
\usage{
addGene(insilicosystem, id = NULL, coding = NULL,
  TargetReaction = NULL, TCrate = NULL, TLrate = NULL,
  RDrate = NULL, PDrate = NULL)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{id}{Integer. The id of the gene to add. Default value: maximum(existing ids) + 1.}

\item{coding}{String. The coding status of the gene (either "PC" or "NC"). If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}

\item{TargetReaction}{String. The biological function of the gene, i.e. the gene expression step targeted by the active product
of the gene.}

\item{TCrate}{Numeric. The transcription rate of the gene. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}

\item{TLrate}{Numeric. The translation rate of the gene. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}

\item{RDrate}{Numeric. The RNA decay rate of the gene. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}

\item{PDrate}{Numeric. The protein decay rate of the gene. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}
}
\value{
The modified in silico system.
}
\description{
Adds a gene in the in silico system with specified parameters if provided, or samples the parameters according to the system parameters.
}
