% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{checkJuliaModules}
\alias{checkJuliaModules}
\title{Checking whether the required Julia modules are installed}
\usage{
checkJuliaModules(silence_warning = FALSE)
}
\arguments{
\item{silence_warning}{Logical, whether the warning issued if required
Julia modules are missing must be silenced. Only for testing. Default
value is FALSE (i.e. by default a warning will be issued if some of the
required modules are missing).}
}
\value{
Nothing
}
\description{
Checking whether the required Julia modules are installed.
}
