% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{checkJuliaVersion}
\alias{checkJuliaVersion}
\title{Checks whether the installed version of Julia is at least a specific version.}
\usage{
checkJuliaVersion(version)
}
\arguments{
\item{version}{Character, on the form "X.Y.Z" where X, Y and Z are integers,
representing the version to compare to.}
}
\value{
\itemize{
\item `-1`: if couldn't check the Julia version
\item `0`: if the current Julia version is lower than the specified `version`
\item `1`: if the current Julia version is at least the specified `version`
}
}
\description{
Compares the installed Julia version to a specific version number.
}
