% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_population.R
\name{createInSilicoPopulation}
\alias{createInSilicoPopulation}
\title{Creates a population of In Silico individuals}
\usage{
createInSilicoPopulation(nInd, insilicosystem, sameInit = F, ...)
}
\arguments{
\item{nInd}{Integer. The number of in silico individuals to create.}

\item{insilicosystem}{An \code{insilicosystem} object. The in silico system from which individuals are created.}

\item{sameInit}{Logical. Do the individuals in the population have the same initial abundance for the different molecules? Default value is \code{FALSE}.}

\item{...}{Other arguments to be passed to the function \code{\link{insilicosystemargs}}.}
}
\value{
An object of class \code{insilicopopulation}, that is a list composed of:
\itemize{
\item \code{GenesVariants} A list of variants segregating in the population for each genes (see \code{\link{createVariants}}).
\item \code{individualsList} A list of In Silico individuals (see \code{\link{createIndividual}}).
\item \code{indargs} An object of class \code{insilicoindividualargs}; the parameters used to create the in silico individuals.
}
}
\description{
Creates a population of in silico individuals to be simulated.
}
