% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{createInSilicoSystem}
\alias{createInSilicoSystem}
\title{Creates an in silico system.}
\usage{
createInSilicoSystem(empty = F, ev = getJuliaEvaluator(), ...)
}
\arguments{
\item{empty}{Logical. Does the regulatory network is empty (= no regulation)? Default value is \code{FALSE}.}

\item{ev}{A Julia evaluator. If none provided select the current evaluator or create one if no evaluator exists.}

\item{...}{Other arguments to be passed to the function \code{\link{insilicosystemargs}}.}
}
\value{
An object of class \code{insilicosystem}, that is a list composed of:
\itemize{
\item \code{genes}: a data-frame of genes (see \code{\link{createGenes}});
\item \code{edg}: a data-frame of edges in the regulatory network (see \code{\link{createMultiOmicNetwork}});
\item \code{mosystem}: a list defining the multi-omic regulatory network (see \code{\link{createMultiOmicNetwork}});
\item \code{sysargs}: An object of class \code{insilicosystemargs}; the parameters used to create the system.
}
}
\description{
Creates an in silico system, i.e. the genes and the regulatory networks defining the system.
}
