% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_population.R
\name{createPopulation}
\alias{createPopulation}
\title{Creates a population of In Silico individuals}
\usage{
createPopulation(nInd, insilicosystem, indargs, sameInit = F)
}
\arguments{
\item{nInd}{Integer. The number of in silico individuals to create.}

\item{insilicosystem}{An \code{insilicosystem} object. The in silico system from which individuals are created.}

\item{indargs}{An object of class \code{insilicoindividualargs} (i.e. a list with parameters for in silico individuals generation).}

\item{sameInit}{Logical. Do the individuals in the population have the same initial abundance for the different molecules? Default value is \code{FALSE}.}
}
\description{
Creates a population of in silico individuals to be simulated.
}
