% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{createRegulatoryNetwork}
\alias{createRegulatoryNetwork}
\title{Creates an in silico regulatory network}
\usage{
createRegulatoryNetwork(regsList, tarsList, reaction, sysargs,
  ev = getJuliaEvaluator())
}
\arguments{
\item{regsList}{A named list of length 2. Element "PC" (resp."NC") is a vector of gene ID of the protein-coding (resp. noncoding) regulators
for the network.}

\item{tarsList}{A named list of length 2. Element "PC" (resp."NC") is a vector of gene ID of the targets of protein-coding (resp. noncoding)
regulators.}

\item{reaction}{String. The ID of the reaction ("TC", "TL", "RD", "PD" or "PTM")}

\item{sysargs}{An object of class \code{insilicosystemargs} (i.e. a list with parameters for in silico system generation).}

\item{ev}{A Julia evaluator. If none provided select the current evaluator or create one if no evaluator exists.}
}
\value{
A list of two elements:
\itemize{
\item \code{edg} a data-frame of edges of the networks;
\item \code{complexes} a list of complexes composition (each element is named with the complex ID, the components are given as gene IDs).
}
The \code{edg} data-frame has the following variables:
\itemize{
\item \code{from} gene ID of the regulator, as a character;
\item \code{to} gene ID of the target, as an integer;
\item \code{TargetReaction} the ID of the reaction (as given by \code{reaction});
\item \code{RegSign} The sign of the reaction ("1" or "-1");
\item \code{RegBy} Is the regulator a protein-coding gene ("PC"), a noncoding gene ("NC") or a complex ("C")?
}
}
\description{
Creates an in silico regulatory network given a list of regulators and targets
}
