% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{createStochSystem}
\alias{createStochSystem}
\title{Creates a stochastic system from an in silico system.}
\usage{
createStochSystem(insilicosystem, indargs, ev = getJuliaEvaluator())
}
\arguments{
\item{insilicosystem}{The in silico system (object of class \code{insilicosystem}, see \code{\link{createInSilicoSystem}}).}

\item{indargs}{An object of class \code{insilicoindividualargs} (i.e. a list with parameters for in silico individuals generation).}

\item{ev}{A Julia evaluator. If none provided select the current evaluator or create one if no evaluator exists.}
}
\value{
A Julia proxy object to retrieve the stochastic system in the Julia evaluator.
}
\description{
Creates a list of molecules, reactions and associated propensities to represent the in silico system.
}
