% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_population.R
\name{createVariants}
\alias{createVariants}
\title{Create variants for genes in the system.}
\usage{
createVariants(genes, indargs)
}
\arguments{
\item{genes}{A data frame of genes in the system (created by the function \code{\link{createGenes}}).}

\item{indargs}{An object of class \code{insilicoindividualargs} (i.e. a list with parameters for in silico individuals generation).}
}
\value{
A list of size G (number of genes in the system) where each element is a matrix corresponding to the QTL effect
coefficients (rows) of the different variants (columns) segregating in the in silico population for the corresponding gene.
}
\description{
Create variants that segregate in the in silico population for each gene in the system.
}
\details{
Returns a list that gives for each gene in the system the existing variants segregating in the in silico population.
A variant is defined by a set of QTL effect coefficients ("qtlTCrate", "qtlRDrate", "qtlTCregbind", "qtlRDregrate",
"qtlactivity", "qtlTLrate", "qtlPDrate", "qtlTLregbind", "qtlPDregrate") that correspond to the impact of genetic
mutations on the variant.
}
