% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{findJuliaNoError}
\alias{findJuliaNoError}
\title{Find a Julia executable}
\usage{
findJuliaNoError(test = FALSE)
}
\arguments{
\item{test}{If TRUE, returns TRUE/FALSE depending on whether or not a Julia executable is found. If FALSE, returns the
path to the Julia executable if it exists.}
}
\value{
TRUE/FALSE or the path to the Julia executable
}
\description{
The function is almost identical to the \code{findJulia} function from the \code{XRJulia} package,
but prevent errors arising from looking for a Julia executable with the "which"/"where" function if Julia
is not present.
}
\examples{
\donttest{
finJuliaNoError(test = T)
}
}
