% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{getComplexes}
\alias{getComplexes}
\title{Retrieves regulatory complexes from in silico system}
\usage{
getComplexes(insilicosystem)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}
}
\value{
A named list, with one element per regulatory complex. Each element is itself a list, with the following components:
\itemize{
\item \code{complexes}: a list of regulatory complexes composition. The names of the elements are the IDs of the complexes, and the
values are vectors of gene IDs constituting each regulatory complex.
\item \code{formationrate}: the formation rate of the complex.
\item \code{dissociationrate}: the dissociation rate of the complex.
}
}
\description{
Returns a list of the regulatory complexes and associated properties in the in silico system.
}
\examples{
\dontrun{
mysystem = createInSilicoSystem(G = 20)
getComplexes(mysystem)
}
}
