% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{getEdges}
\alias{getEdges}
\title{Retrieves regulatory edges from in silico system}
\usage{
getEdges(insilicosystem)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}
}
\value{
A data-frame with the following columns:
\itemize{
\item \code{from}: gene ID of the regulator, as a character;
\item \code{to}: gene ID of the target, as an integer;
\item \code{TargetReaction}: the ID of the reaction (as given by \code{reaction});
\item \code{RegSign}: The sign of the reaction ("1" or "-1");
\item \code{RegBy}: Is the regulator a protein-coding gene ("PC"), a noncoding gene ("NC") or a complex ("C")?
}
}
\description{
Returns a data-frame of the regulatory edges and associated properties in the in silico system.
}
\examples{
\dontrun{
mysystem = createInSilicoSystem(G = 10)
getGenes(mysystem)
}
}
