% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{getGenes}
\alias{getGenes}
\title{Retrieve genes from in silico system}
\usage{
getGenes(insilicosystem)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}
}
\value{
A data-frame with the following columns:
\itemize{
\item \code{id}: Integer, ID of the genes;
\item \code{coding}: coding status of the genes (either "PC" for protein-coding or "NC" for noncoding);
\item \code{TargetReaction}: the biological function of the genes ("TC": transcription regulator, "TL": translation regulator, "RD": RNA decay
regulator, "PD": protein decay regulator, "PTM": post-translational modification regulator, "MR": metabolic enzyme).;
\item \code{PTMform}: Does the gene have a PTM form?;
\item \code{Active form}: what is the active form of the gene? "R" for noncoding genes, "P" for protein-coding genes,
"Pm" for protein-coding genes with a PTM form;
\item \code{TCrate}: transcription rate of the genes;
\item \code{TLrate}: translation rate of the genes;
\item \code{RDrate}: RNA decay rate of the genes;
\item \code{PDrate}: Protein decay rate of the genes.
}
}
\description{
Returns a data-frame of the genes and associated properties in the in silico system.
}
\examples{
\dontrun{
mysystem = createInSilicoSystem(G = 10)
getGenes(mysystem)
}
}
