% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{getJuliaEvaluator}
\alias{getJuliaEvaluator}
\title{Returns the current Julia evaluator.}
\usage{
getJuliaEvaluator()
}
\value{
A Julia evaluator from \code{XRJulia} package.
}
\description{
Returns the current Julia evaluator; if none, starts a new one.
}
\details{
\code{getJuliaEvaluator} is similar to the \code{XRJulia} function \code{RJulia},
but if no evaluator exists, creates a new one and loads \code{sismonr} Julia functions
on it.
}
\examples{
\donttest{
getJuliaEvaluator()
}
}
