% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{getReactions}
\alias{getReactions}
\title{Retrieves the stochastic model for an in silico system}
\usage{
getReactions(insilicosystem, insilicopopulation = NULL)
}
\arguments{
\item{insilicosystem}{The in silico system to be simulated (see \code{\link{createInSilicoSystem}}).}

\item{insilicopopulation}{Optional, the in silico population to be simulated (see \code{\link{createInSilicoPopulation}}).}
}
\value{
A data-frame, with the following columns:
\itemize{
\item \code{reaction}: the stochastic reaction;
\item \code{reaction_name}: the name of the stochastic reaction;
\item \code{reaction_rate}: the formula to compute the reaction rate, as a character.
}
In addition, if an in silico population is provided, the data-frame will include one column per in silico individual with the rate of the reactions
for the individual.
}
\description{
Returns a data-frame with the list of reactions with associated name and rate in the stochastic model generated from the in silico system. If
an in silico population is provided, the rate of each reaction will be calculated for each in silico individual in the population.
}
\examples{
\dontrun{
mysystem = createInSilicoSystem(G = 5)
getReactions(mysystem)
}
}
