% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_classes.R
\name{insilicosystemargs}
\alias{insilicosystemargs}
\title{Constructor function for the \code{insilicosystemsargs} class.}
\usage{
insilicosystemargs(G = 10, PC.p = 0.7, PC.TC.p = 0.4,
  PC.TL.p = 0.3, PC.RD.p = 0.1, PC.PD.p = 0.1, PC.PTM.p = 0.05,
  PC.MR.p = 0.05, NC.TC.p = 0.3, NC.TL.p = 0.3, NC.RD.p = 0.3,
  NC.PD.p = 0.05, NC.PTM.p = 0.05, TC.pos.p = 0.5, TL.pos.p = 0.5,
  PTM.pos.p = 0.5, basal_transcription_rate_samplingfct = function(x) {
      logval = rnorm(x, mean = 0.5, sd = 0.5)     val = 10^logval    
  return(val/3600) }, basal_translation_rate_samplingfct = function(x) { 
     logval = rnorm(x, mean = 2.5, sd = 0.8)     val = 10^logval    
  return(val/3600) }, basal_RNAlifetime_samplingfct = function(x) {    
  logval = rnorm(x, mean = 0, sd = 0.5)     val = 10^logval     return(val
  * 3600) }, basal_protlifetime_samplingfct = function(x) {     logval =
  rnorm(x, mean = 1.75, sd = 0.5)     val = 10^logval     return(val *
  3600) }, TC.PC.outdeg.distr = "powerlaw",
  TC.NC.outdeg.distr = "powerlaw", TC.PC.outdeg.exp = 3,
  TC.NC.outdeg.exp = 1, TC.PC.indeg.distr = "powerlaw",
  TC.NC.indeg.distr = "powerlaw", TC.PC.autoregproba = 0.2,
  TC.NC.autoregproba = 0, TC.PC.twonodesloop = FALSE,
  TC.NC.twonodesloop = FALSE, TCbindingrate_samplingfct = function(x) {
      runif(x, 0.001, 0.01) }, TCunbindingrate_samplingfct = function(x)
  {     runif(x, 0.001, 0.01) }, TCfoldchange_samplingfct = function(x) {
      truncnorm::rtruncnorm(x, a = 1, mean = 3, sd = 10) },
  TL.PC.outdeg.distr = "powerlaw", TL.NC.outdeg.distr = "powerlaw",
  TL.PC.outdeg.exp = 3, TL.NC.outdeg.exp = 1,
  TL.PC.indeg.distr = "powerlaw", TL.NC.indeg.distr = "powerlaw",
  TL.PC.autoregproba = 0.2, TL.NC.autoregproba = 0,
  TL.PC.twonodesloop = FALSE, TL.NC.twonodesloop = FALSE,
  TLbindingrate_samplingfct = function(x) {     runif(x, 0.001, 0.01) },
  TLunbindingrate_samplingfct = function(x) {     runif(x, 0.001, 0.01)
  }, TLfoldchange_samplingfct = function(x) {     sample(2:30, x, replace
  = T) }, RD.PC.outdeg.distr = "powerlaw",
  RD.NC.outdeg.distr = "powerlaw", RD.PC.outdeg.exp = 3,
  RD.NC.outdeg.exp = 1, RD.PC.indeg.distr = "powerlaw",
  RD.NC.indeg.distr = "powerlaw", RD.PC.autoregproba = 0.2,
  RD.NC.autoregproba = 0, RD.PC.twonodesloop = FALSE,
  RD.NC.twonodesloop = FALSE, RDregrate_samplingfct = function(x) {    
  runif(x, 0.001, 0.01) }, PD.PC.outdeg.distr = "powerlaw",
  PD.NC.outdeg.distr = "powerlaw", PD.PC.outdeg.exp = 3,
  PD.NC.outdeg.exp = 1, PD.PC.indeg.distr = "powerlaw",
  PD.NC.indeg.distr = "powerlaw", PD.PC.autoregproba = 0.2,
  PD.NC.autoregproba = 0, PD.PC.twonodesloop = FALSE,
  PD.NC.twonodesloop = FALSE, PDregrate_samplingfct = function(x) {    
  runif(x, 0.001, 0.01) }, PTM.PC.outdeg.distr = "powerlaw",
  PTM.NC.outdeg.distr = "powerlaw", PTM.PC.outdeg.exp = 3,
  PTM.NC.outdeg.exp = 1, PTM.PC.indeg.distr = "powerlaw",
  PTM.NC.indeg.distr = "powerlaw", PTM.PC.autoregproba = 0.2,
  PTM.NC.autoregproba = 0, PTM.PC.twonodesloop = FALSE,
  PTM.NC.twonodesloop = FALSE, PTMregrate_samplingfct = function(x) {  
    runif(x, 0.001, 0.01) }, regcomplexes = "prot",
  regcomplexes.p = 0.3, regcomplexes.size = 2,
  complexesformationrate_samplingfct = function(x) {     runif(x, 0.001,
  0.01) }, complexesdissociationrate_samplingfct = function(x) {    
  runif(x, 0.001, 0.01) }, mycolsCS = c(PC = "#e03616", NC = "#58355e",
  Tot = "#31161F"), mycolsGF = c(TC = "#FF7F11", TL = "#FF963C", RD =
  "#5AB7A4", PD = "#78C4B4", PTM = "#FF1B1C", MR = "#FF6D6E"),
  mycolsPosNeg = c(`1` = "#D63230", `-1` = "#69BAF4"))
}
\arguments{
\item{G}{Integer. Number of genes in the system. Default value is 10.}

\item{PC.p}{Numeric. Probability of each gene to be a protein-coding gene. Default value is 0.7.}

\item{PC.TC.p}{Numeric. Ratio of regulators of transcription among the protein-coding genes. Default value is 0.4.}

\item{PC.TL.p}{Numeric. Ratio of regulators of translation among the protein-coding genes. Default value is 0.3.}

\item{PC.RD.p}{Numeric. Ratio of regulators of RNA decay among the protein-coding genes. Default value is 0.1.}

\item{PC.PD.p}{Numeric. Ratio of regulators of protein decay among the protein-coding genes. Default value is 0.1.}

\item{PC.PTM.p}{Numeric. Ratio of regulators of protein post-translational modification among the protein-coding genes. Default value is 0.05.}

\item{PC.MR.p}{Numeric. Ratio of metabolic enzymes among the protein-coding genes. Default value is 0.05.}

\item{NC.TC.p}{Numeric. Ratio of regulators of transcription among the noncoding genes. Default value is 0.3.}

\item{NC.TL.p}{Numeric. Ratio of regulators of translation among the noncoding genes. Default value is 0.3.}

\item{NC.RD.p}{Numeric. Ratio of regulators of RNA decay among the noncoding genes. Default value is 0.3.}

\item{NC.PD.p}{Numeric. Ratio of regulators of protein decay among the noncoding genes. Default value is 0.05.}

\item{NC.PTM.p}{Numeric. Ratio of regulators of protein post-translational modification among the noncoding genes. Default value is 0.05.}

\item{TC.pos.p}{Numeric. Probability that the transcription is positively regulated by regulators. Default value is 0.5.}

\item{TL.pos.p}{Numeric. Probability that the translation is positively regulated by regulators. Default value is 0.5.}

\item{PTM.pos.p}{Numeric. Probability that the regulators transform the original protein into its modified form
(as opposed to transforming the modified protein back into its original form). Default value is 0.5.}

\item{basal_transcription_rate_samplingfct}{Function from which the transcription rates of genes are sampled (input x is the required sample size). Default value is
Values from Schwanhausser et al., 2013: transcription rate distribution log-normal, from 0.1 to 100 mRNA/hour -> we want the transcription rate in seconds. Default value is}

\item{basal_translation_rate_samplingfct}{Function from which the translation rates of genes are sampled (input x is the required sample size). Default value is
Values from Schwanhausser et al., 2013: translation rate distribution log-normal, from 0.1 to 10^5 protein/mRNA/hour -> we want the transcription rate in seconds. Default value is}

\item{basal_RNAlifetime_samplingfct}{Function from which the transcript lifetimes are sampled (input x is the required sample size). Default value is
Values from Schwanhausser et al., 2013: RNAs half-life distribution log-normal (chose boundary values to be from 1 min to 100 hours) -> we want the half-life in seconds. Default value is}

\item{basal_protlifetime_samplingfct}{Function from which the protein lifetime are sampled (input x is the required sample size). Default value is
Values from Schwanhausser et al., 2013: proteins half-life distribution log-normal (chose boundary values to be from 1 hour to 1000 hours) -> we want the half-life in seconds. Default value is}

\item{TC.PC.outdeg.distr}{Form of the distribution of the number of targets of transcription factors; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TC.NC.outdeg.distr}{Form of the distribution of the number of targets of noncoding RNAs regulating transcription; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TC.PC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the protein regulators in the transcription regulation graph. Default value is 3.}

\item{TC.NC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the noncoding regulators in the transcription regulation graph. Default value is 1.}

\item{TC.PC.indeg.distr}{Type of preferential attachment for the targets of protein regulators in the transcription regulation graph; can be either "powerlaw" or "exponential"). Default value is "powerlaw".}

\item{TC.NC.indeg.distr}{Type of preferential attachment for the targets of ncRNAs in the transcription regulation graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TC.PC.autoregproba}{Numeric. Probability of protein regulators to perform autoregulation in the transcription regulation graph. Default value is 0.2.}

\item{TC.NC.autoregproba}{Numeric. Probability of ncRNAs to perform autoregulation in the transcription regulation graph. Default value is 0.}

\item{TC.PC.twonodesloop}{Logical. Are 2-nodes loops authorised in the transcription regulation graph with protein regulators? Default value is FALSE.}

\item{TC.NC.twonodesloop}{Logical. Are 2-nodes loops authorised in the transcription regulation graph with noncoding regulators? Default value is FALSE.}

\item{TCbindingrate_samplingfct}{Function from which the binding rate of transcription regulators on target are sampled (input x is the required sample size). Default value is}

\item{TCunbindingrate_samplingfct}{Function from which the unbinding rate of transcription regulators from target are sampled (input x is the required sample size). Default value is}

\item{TCfoldchange_samplingfct}{Function from which the transcription fold change induced by a bound regulator are sampled (input x is the required sample size). Default value is}

\item{TL.PC.outdeg.distr}{Form of the distribution of the number of targets of translation factors; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TL.NC.outdeg.distr}{Form of the the distribution of the number of targets of noncoding RNAs regulating translation; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TL.PC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the protein regulators in the translation regulation graph. Default value is 3.}

\item{TL.NC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the noncoding regulators in the translation regulation graph. Default value is 1.}

\item{TL.PC.indeg.distr}{Type of preferential attachment for the targets of protein regulators in the translation regulation graph can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TL.NC.indeg.distr}{Type of preferential attachment for the targets of ncRNAs in the translation regulation graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{TL.PC.autoregproba}{Numeric. Probability of protein regulators to perform autoregulation in the translation regulation graph. Default value is 0.2.}

\item{TL.NC.autoregproba}{Numeric. Probability of ncRNAs to perform autoregulation in the translation regulation graph. Default value is 0.}

\item{TL.PC.twonodesloop}{Logical. Are 2-nodes loops authorised in the translation regulation graph with protein regulators? Default value is FALSE.}

\item{TL.NC.twonodesloop}{Logical. Are 2-nodes loops authorised in the translation regulation graph with noncoding regulators? Default value is FALSE.}

\item{TLbindingrate_samplingfct}{Function from which the binding rate of translation regulators on target are sampled (input x is the required sample size). Default value is}

\item{TLunbindingrate_samplingfct}{Function from which the unbinding rate of translation regulators from target are sampled (input x is the required sample size). Default value is}

\item{TLfoldchange_samplingfct}{Function from which the translation fold change induced by a bound regulator are sampled (input x is the required sample size). Default value is}

\item{RD.PC.outdeg.distr}{Form of the distribution of the number of targets of RNA decay factors; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{RD.NC.outdeg.distr}{Form of the the distribution of the number of targets of noncoding RNAs regulating RNA decay; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{RD.PC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the protein regulators in the RNA decay regulation graph. Default value is 3.}

\item{RD.NC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the noncoding regulators in the RNA decay regulation graph. Default value is 1.}

\item{RD.PC.indeg.distr}{Type of preferential attachment for the targets of protein regulators in the RNA decay graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{RD.NC.indeg.distr}{Type of preferential attachment for the targets of ncRNAs in the RNA decay graph;can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{RD.PC.autoregproba}{Numeric. Probability of protein regulators to perform autoregulation in the RNA decay regulation graph. Default value is 0.2.}

\item{RD.NC.autoregproba}{Numeric. Probability of ncRNAs to perform autoregulation in the RNA decay regulation graph. Default value is 0.}

\item{RD.PC.twonodesloop}{Logical. Are 2-nodes loops authorised in the RNA decay regulation graph with protein regulators? Default value is FALSE.}

\item{RD.NC.twonodesloop}{Logical. Are 2-nodes loops authorised in the RNA decay regulation graph with noncoding regulators? Default value is FALSE.}

\item{RDregrate_samplingfct}{Function from which the RNA decay rates of targets of RNA decay regulators are sampled (input x is the required sample size). Default value is}

\item{PD.PC.outdeg.distr}{Form of the distribution of the number of targets of protein decay factors; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PD.NC.outdeg.distr}{Form of the the distribution of the number of targets of noncoding RNAs regulating protein decay; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PD.PC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the protein regulators in the protein decay regulation graph. Default value is 3.}

\item{PD.NC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the noncoding regulators in the protein decay regulation graph. Default value is 1.}

\item{PD.PC.indeg.distr}{Type of preferential attachment for the targets of protein regulators in the protein decay regulation graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PD.NC.indeg.distr}{Type of preferential attachment for the targets of ncRNAs in the protein decay graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PD.PC.autoregproba}{Numeric. Probability of protein regulators to perform autoregulation in the protein decay regulation graph. Default value is 0.2.}

\item{PD.NC.autoregproba}{Numeric. Probability of ncRNAs to perform autoregulation in the protein decay regulation graph. Default value is 0.}

\item{PD.PC.twonodesloop}{Logical. Are 2-nodes loops authorised in the protein decay graph with protein regulators in the protein decay regulation graph? Default value is FALSE.}

\item{PD.NC.twonodesloop}{Logical. Are 2-nodes loops authorised in the protein decay graph with noncoding regulators in the protein decay regulation graph? Default value is FALSE.}

\item{PDregrate_samplingfct}{Function from which the protein decay rates of targets of protein decay regulators are sampled (input x is the required sample size). Default value is}

\item{PTM.PC.outdeg.distr}{Form of the distribution of the number of targets of protein post-translational modification regulators; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PTM.NC.outdeg.distr}{Form of the the distribution of the number of targets of noncoding RNAs regulating protein post-translational modification; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PTM.PC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the protein regulators in the protein post-translational modification graph. Default value is 3.}

\item{PTM.NC.outdeg.exp}{Numeric. Exponent of the distribution for the out-degree of the noncoding regulators in the protein post-translational modification graph. Default value is 1.}

\item{PTM.PC.indeg.distr}{Type of preferential attachment for the targets of protein regulators in the protein post-translational modification graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PTM.NC.indeg.distr}{Type of preferential attachment for the targets of ncRNAs in the protein post-translational modification graph; can be either "powerlaw" or "exponential". Default value is "powerlaw".}

\item{PTM.PC.autoregproba}{Numeric. Probability of protein regulators to perform autoregulation. Default value is 0.2.}

\item{PTM.NC.autoregproba}{Numeric. Probability of ncRNAs to perform autoregulation. Default value is 0.}

\item{PTM.PC.twonodesloop}{Logical. Are 2-nodes loops authorised in the protein post-translational modification graph with protein regulators? Default value is FALSE.}

\item{PTM.NC.twonodesloop}{Logical. Are 2-nodes loops authorised in the protein post-translational modification graph with noncoding regulators? Default value is FALSE.}

\item{PTMregrate_samplingfct}{Function from which the protein transformation rates of targets of post-translational modification regulators are sampled (input x is the required sample size). Default value is}

\item{regcomplexes}{Can the regulators controlling a common target form regulatory complexes in the protein post-translational modification graph? Can be 'none', 'prot' (only protein can form regulatory complexes) or 'both'
(both RNAs and proteins can form regulatory complexes). Default value is "prot".}

\item{regcomplexes.p}{Numeric. Probability that regulators controlling a common target form regulatory complexes; ignore if regcomplexes = 'none'. Default value is 0.3.}

\item{regcomplexes.size}{Integer. Number of components of a regulatory complex; ignore if regcomplexes = 'none'. Default value is 2.}

\item{complexesformationrate_samplingfct}{Function from which the formation rate of regulatory complexes are sampled (input x is the required sample size). Default value is}

\item{complexesdissociationrate_samplingfct}{Function from which the dissociation rate of regulatory complexes are sampled (input x is the required sample size). Default value is}

\item{mycolsCS}{Named vector of colour names or code. Colours used in the plots to represent portein-coding and noncoding genes. Default value is c("PC" = "#e03616",  "NC" = "#58355e", "Tot" = "#31161F").}

\item{mycolsGF}{Colours used in the plots to represent the different gene expression steps (transcription, translation, etc). Default value is c("TC" = "#FF7F11", "TL" = "#FF963C", "RD" = "#5AB7A4", "PD" = "#78C4B4", "PTM" = "#FF1B1C", "MR" = "#FF6D6E").}

\item{mycolsPosNeg}{Colours used in plots to represent positive/negative regulatory interactions. Default value is c("1" = "#D63230", "-1" = "#69BAF4").}
}
\value{
An object of the class \code{insilicosystemargs}, that is a list of the different parameters.
}
\description{
Constructor function for the \code{insilicosystemsargs} class, with default values if not provided by the user.
}
