% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{installJuliaModules}
\alias{installJuliaModules}
\title{Install required Julia modules}
\usage{
installJuliaModules()
}
\value{
Nothing.
}
\description{
Installed required Julia modules. Requires user permission before installing the modules.
}
