% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{mergeAlleleAbundance}
\alias{mergeAlleleAbundance}
\title{Merge the different allelic versions of the molecules.}
\usage{
mergeAlleleAbundance(simulation)
}
\arguments{
\item{simulation}{The result of \code{\link{simulateInSilicoSystem}} or \code{\link{simulateParallelInSilicoSystem}}, that is
a list in which one element named "Simulation" is a dataframe with the abundance of the different molecules over time}
}
\value{
A dataframe in which the abundance of the different allelic versions of the same molecule have been added.
}
\description{
Sum the abundance of the different allelic versions of each molecule.
}
