% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{mergeComplexesAbundance}
\alias{mergeComplexesAbundance}
\title{Merge the free and in complexe versions of molecules.}
\usage{
mergeComplexesAbundance(df)
}
\arguments{
\item{df}{A dataframe with the abundance of the different molecules over time (from \code{\link{simulateInSilicoSystem}}
or \code{\link{simulateParallelInSilicoSystem}}).}
}
\value{
A dataframe in which the abundance of free and in complexe versions of a molecule have been added.
}
\description{
Sum the abundance of the free and in complexe versions of each molecule.
}
