% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{newJuliaEvaluator}
\alias{newJuliaEvaluator}
\title{Creates a new ready-to-use Julia evaluator.}
\usage{
newJuliaEvaluator(port = NULL)
}
\arguments{
\item{port}{An integer specifying the port to be used. Default \code{NULL}.}
}
\value{
A Julia Evaluator from the XRJulia package.
}
\description{
\code{newJuliaEvaluator} opens a new Julia evaluator and loads the required functions on it.
}
\details{
If no port number is specified, the \code{RJulia} function chooses a port to open the evaluator.
}
\examples{
\donttest{
ev = newJuliaEvaluator()
}
}
