% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{plotSimulation}
\alias{plotSimulation}
\title{Plots the result of a simulation}
\usage{
plotSimulation(simdf, inds, mergeAllele = T, mergePTM = T,
  mergeComplexes = F, yLogScale = T, nIndPerRow = 3, ...)
}
\arguments{
\item{simdf}{The dataframe with the result of the simulation (see \code{\link{simulateInSilicoSystem}})}

\item{inds}{A vector of in silico individual names for which to plot the expression profiles}

\item{mergeAllele}{Are the gene products originating from different alleles merged? Default TRUE. Also see \code{\link{mergeAlleleAbundance}}}

\item{mergePTM}{Are the modified and non-modified versions of the proteins merged? Default TRUE. Also see \code{\link{mergePTMAbundance}}}

\item{mergeComplexes}{Are the free and in complex gene products merged? Default FALSE Also see \code{\link{mergeComplexesAbundance}}}

\item{yLogScale}{Is the y-axis of the plot in log10-scale? If so, the abundance of each species at each time-point is increased by 1 to avoid zero values. Default TRUE.}

\item{nIndPerRow}{Positive integer, the number of individuals to plot per row. Default 3.}

\item{...}{Any additional parameter to be passed to \code{\link[ggplot2]{theme}} for the plot of each individual.}
}
\value{
A plot from \code{\link[ggpubr]{ggarrange}}.
}
\description{
Automatically plots the result of a simulation for the selected in silico individuals.
}
\examples{
\donttest{
mysystem = createInSilicoSystem(G = 5, regcomplexes = "none")
mypop = createInSilicoPopulation(15, mysystem, ploidy = 2)
sim = simulateParallelInSilicoSystem(mysystem, mypop, 100, ntrials = 5)
plotSimulation(sim$Simulation, c("Ind1", "Ind2", "Ind3", "Ind4"),
 axis.title = element_text(color = "red"))
}
}
