% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{remove_julia_evaluator}
\alias{remove_julia_evaluator}
\title{Closes a Julia evaluator.}
\usage{
remove_julia_evaluator(ev)
}
\arguments{
\item{ev}{A Julia evaluator from the XRJulia package.}
}
\description{
\code{remove_julia_evaluator} closes a Julia evaluator from XRJulia package.
}
