% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{removeComplex}
\alias{removeComplex}
\title{Removes a regulatory complex from the in silico system.}
\usage{
removeComplex(insilicosystem, name)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{name}{String. The name of the regulatory complex to remove.}
}
\value{
The modified in silico system.
}
\description{
Removes a regulatory complex from the in silico system. Any edge involving this complex is removed from the system.
}
