% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{removeEdg}
\alias{removeEdg}
\title{Removes an edge from the in silico system.}
\usage{
removeEdg(insilicosystem, regID, tarID)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{regID}{Integer. The ID of the regulator gene.}

\item{tarID}{Integer. The ID of the target gene.}
}
\value{
The modified in silico system.
}
\description{
Removes an edge in the in silico system between specified genes.
}
