% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_evaluators.R
\name{removeJuliaEvaluator}
\alias{removeJuliaEvaluator}
\title{Closes a Julia evaluator.}
\usage{
removeJuliaEvaluator(ev)
}
\arguments{
\item{ev}{A Julia evaluator from the XRJulia package.}
}
\description{
\code{removeJuliaEvaluator} closes a Julia evaluator from XRJulia package.
}
\examples{
\donttest{
removeJuliaEvaluator(getJuliaEvaluator())
}
}
