% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{addComplex}
\alias{addComplex}
\title{Adds a regulatory complex in the in silico system.}
\usage{
addComplex(insilicosystem, compo, formationrate = NULL,
  dissociationrate = NULL)
}
\arguments{
\item{insilicosystem}{The in silico system (see \code{\link{createInSilicoSystem}}).}

\item{compo}{An integer vector, which each element being the ID of the genes composing the complex. Must be of the same length as
specified in the system arguments (i.e. regcomplexes.size, see \code{\link{insilicosystemsargs}}). All genes composing the compex must
have the same biological function (i.e. same "TargetReaction" parameter).}

\item{formationrate}{The formation rate of the complex. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}

\item{dissociationrate}{The dissociation rate of the complex. If none provided, randomly chosen according to the
parameters provided in sysargs (see \code{\link{insilicosystemsargs}}).}
}
\value{
The modified in silico system.
}
\description{
Adds a regulatory complex in the in silico system with specified parameters if provided, or samples the parameters according to the system parameters.
}
