% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{createEmptyMultiOmicNetwork}
\alias{createEmptyMultiOmicNetwork}
\title{Creates an empty in silico multi omic network}
\usage{
createEmptyMultiOmicNetwork(genes)
}
\arguments{
\item{genes}{A data-frame of genes in the system (see \code{\link{createGenes}}).}
}
\value{
An in silico multi omic network, that is a list of:
\itemize{
\item \code{genes}: the modified data-frame of genes;
\item \code{edg}: A data-frame of edges in the multi-omic network (see details below).
\item \code{mosystem}: A list of the different regulation networks in the system and associated information.
}
The \code{mosystem} list is composed of:
\itemize{
\item \code{TCRN_edg}, \code{TLRN_edg}, \code{RDRN_edg}, \code{PDRN_edg}, \code{PTMRN_edg}: data-frames of edges for the different
regulatory networks, with in addition to the usual fields in the edg data frame contains columns for kinetic parameters of the
regulation. All empty.
\item \code{complexes}: a list of regulatory complexes composition. The names of the elements are the IDs of the complexes, and the
values are vectors of gene IDs constituting each regulatory complex. Empty list.
\item \code{complexeskinetics}: a list of regulatory complexes kinetic parameters. Empty list.
}
}
\description{
Creates an empty in silico multi omic network from a data-frame of genes in the system (no regulatory interactions) (cf
\code{\link{createMultiOmicNetwork}}).
}
