% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_population.R
\name{createIndividual}
\alias{createIndividual}
\title{Create an In Silico individual}
\usage{
createIndividual(variantsList, indargs, sameInit = F)
}
\arguments{
\item{variantsList}{List of segregating variants (created by \code{\link{createVariants}}). Each element is a matrix of the existing variants
for a given gene in the system.}

\item{indargs}{An object of class \code{insilicoindividualargs} (i.e. a list with parameters for in silico individuals generation).}

\item{sameInit}{Boolean. Does the individual have identical initial abundance to the rest of the population?}
}
\value{
An object of class \code{insilicoindividual}.
}
\description{
Creates a in silico individual to be simulated (object of class \code{insilicoindividual}).
}
