% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_system.R
\name{createMultiOmicNetwork}
\alias{createMultiOmicNetwork}
\title{Creates an in silico multi omic network}
\usage{
createMultiOmicNetwork(genes, sysargs, ev = getJuliaEvaluator())
}
\arguments{
\item{genes}{A data-frame of genes in the system (see \code{\link{createGenes}}).}

\item{sysargs}{An object of class \code{insilicosystemargs} (i.e. a list with parameters for in silico system generation).}

\item{ev}{A Julia evaluator. If none provided select the current evaluator or create one if no evaluator exists.}
}
\value{
An in silico multi omic network, that is a list of:
\itemize{
\item \code{genes} the modified data-frame of genes;
\item \code{edg} A data-frame of edges in the multi-omic network (see details below).
\item \code{mosystem} A list of the different regulation networks in the system and associated information.
}
The \code{mosystem} list is composed of:
\code{TCRN_edg}, \code{TLRN_edg}, \code{RDRN_edg}, \code{PDRN_edg}, \code{PTMRN_edg}: data-frames of edges for the different
regulatory networks, with in addition to the usual fields in the edg data frame contains columns for kinetic parameters of the
regulation.
\code{complexes}: a list of regulatory complexes composition. The names of the elements are the IDs of the complexes, and the
values are vectors of gene IDs constituting each regulatory complex.
\code{complexeskinetics}: a list of regulatory complexes kinetic parameters.
}
\description{
Creates an in silico multi omic network from a data-frame of genes in the system.
}
