% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{df2list}
\alias{df2list}
\title{Tranforms a data-frame into a list.}
\usage{
df2list(mydf)
}
\arguments{
\item{mydf}{A data-frame.}
}
\value{
A named list with elements corresponding to the columns of the input data-frame.
}
\description{
Transforms a data-frame into a list. The elements of the list correspond to
the columns of the data-frame.
}
