% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{mergeAlleleAbundance}
\alias{mergeAlleleAbundance}
\title{Merge the different allelic versions of the molecules.}
\usage{
mergeAlleleAbundance(df)
}
\arguments{
\item{df}{A dataframe with the abundance of the different molecules over time (from \code{\link{simulateInSilicoSystem}}
or \code{\link{simulateParallelInSilicoSystem}}).}
}
\value{
A dataframe in which the abundance of the different allelic versions of the same molecule have been added.
}
\description{
Sum the abundance of the different allelic versions of each molecule.
}
