% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{mergePTMAbundance}
\alias{mergePTMAbundance}
\title{Merge the original and PTM versions of the proteins.}
\usage{
mergePTMAbundance(df)
}
\arguments{
\item{df}{A dataframe with the abundance of the different molecules over time (from \code{\link{simulateInSilicoSystem}}
or \code{\link{simulateParallelInSilicoSystem}}).}
}
\value{
A dataframe in which the abundance of original and modified versions of a protein have been added.
}
\description{
Sum the abundance of the original and modified (PTM) versions of each protein.
}
