% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{simulateInSilicoSystem}
\alias{simulateInSilicoSystem}
\title{Simulates a in-silico system}
\usage{
simulateInSilicoSystem(insilicosystem, insilicopopulation, simtime,
  nepochs = -1, ntrialsPerInd = 1, simalgorithm = "Direct",
  writefile = T, filepath = getwd(), filename = "simulation",
  ev = getJuliaEvaluator())
}
\arguments{
\item{insilicosystem}{The in silico system to be simulated (see \code{\link{createInSilicoSystem}}).}

\item{insilicopopulation}{The in silico insilicopopulation to be simulated (see \code{\link{createInSilicoPopulation}}).}

\item{simtime}{The amount of time to simulate the model (in seconds).}

\item{nepochs}{The number of times to record the state of the system during the simulation.}

\item{ntrialsPerInd}{The number of times the simulation must be replicated.}

\item{simalgorithm}{The name of the simulation algorithm to use in the Julia function \code{simulate} from the module \code{BioSimulator}.
Can be one of "Direct", "FirstReaction", "NextReaction", "OptimizedDirect", "TauLeaping", "StepAnticipation".}

\item{writefile}{Does the julia function write the species and reactions lists in a text file?}

\item{filepath}{If writefile = \code{TRUE}, path to the folder in the which the files will be created.}

\item{filename}{If writefile = \code{TRUE}, prefix of the files created to store the lists of species and reactions.}

\item{ev}{A Julia evaluator. If none provided select the current evaluator or create one if no evaluator exists.}
}
\value{
A list composed of:
\itemize{
\item \code{resTable}: A list where each element is the data-frame of simulated expression profiles for an individual in the in silico population.
\item \code{runningtime}: A vector of running time of all simulations
\item \code{stochmodel}: A Julia proxy object to retrieve the stochastic system in the Julia evaluator.
}
}
\description{
Simulates (stochastically) the behaviour of an in silico system over time
}
